MASTER: 05-simple-math : scala : support decimals

+++ 05-simple-math/scala-simple-math/src/main/scala/com/lochnesh/exercises/simplemath/Computations.scala
@@ -2,9 +2,9 @@ package com.lochnesh.exercises.simplemath
 
 class Computations {
 
-  def add(firstNumber: Int, secondNumber: Int) = firstNumber + secondNumber
-  def subtract(firstNumber: Int, secondNumber: Int) = firstNumber - secondNumber
-  def multiply(firstNumber: Int, secondNumber: Int) = firstNumber * secondNumber
-  def divide(firstNumber: Int, secondNumber: Int) = firstNumber / secondNumber
+  def add(firstNumber: BigDecimal, secondNumber: BigDecimal) = firstNumber + secondNumber
+  def subtract(firstNumber: BigDecimal, secondNumber: BigDecimal) = firstNumber - secondNumber
+  def multiply(firstNumber: BigDecimal, secondNumber: BigDecimal) = firstNumber * secondNumber
+  def divide(firstNumber: BigDecimal, secondNumber: BigDecimal) = firstNumber / secondNumber
 
 }
